Loop only through subdirectories but not files in Plugins directory.

+++ CleanAuxFiles.py
@@ -22,11 +22,12 @@ for name in os.listdir(folder_path):
         print(f"Checking {fullname}")
         for plugin_name in os.listdir(fullname):
             plugin_path = os.path.join(fullname, plugin_name)
-            for plugin_folder in os.listdir(plugin_path):
-                full_plugin_folder = os.path.join(plugin_path, plugin_folder)
-                if plugin_folder in subfolder_names and os.path.isdir(full_plugin_folder):
-                    print(f"Deleting {plugin_path}")
-                    shutil.rmtree(full_plugin_folder)
+            if os.path.isdir(plugin_path):
+                for plugin_folder in os.listdir(plugin_path):
+                    full_plugin_folder = os.path.join(plugin_path, plugin_folder)
+                    if plugin_folder in subfolder_names and os.path.isdir(full_plugin_folder):
+                        print(f"Deleting {plugin_path}")
+                        shutil.rmtree(full_plugin_folder)
 
     if os.path.splitext(name)[-1] in file_extension:
         print(f"Deleting {fullname}")
